pyproject.toml
[project]
name = "poll-creation-tg-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-generativeai>=0.8.6",
    "python-dotenv>=1.2.2",
    "python-telegram-bot[job-queue]>=22.7",
    "sqlalchemy>=2.0.49",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
